prediction/requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
numpy==2.0.1
fastapi==0.115.2
uvicorn==0.30.6
requests==2.32.3
urllib3==1.26.18
psycopg2-binary==2.9.9
